Enhanced environment variables' support.
Previously only passwords (for "ssh" connections and "sudo")
were able to be configured using environment variables.
Now even more configurations can be done this way, which will make it easier
to run the same flows on another host (or just in case of the IP change).
The following fields now support environment variables:
1. Remote (ssh) host.
2. Remote (ssh) user.
3. Privileged (sudo) user.

Signed-off-by: Pavel Bar <pbar@redhat.com>